+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-type-select/src/main/java/com/liferay/dynamic/data/mapping/type/select/internal/SelectDDMFormFieldTypeSettings.java
@@ -107,7 +107,7 @@ public interface SelectDDMFormFieldTypeSettings
 
 	@DDMFormField(
 		label = "%allow-multiple-selections",
-		properties = {"showAsSwitcher=true"}
+		properties = {"showAsSwitcher=true"}, visibilityExpression = "FALSE"
 	)
 	public boolean multiple();