+++ modules/sdk/gradle-plugins-node/CHANGELOG.markdown
@@ -77,6 +77,13 @@ and the `npm-shrinkwrap.json` file from the project, if present.
 - [LPS-69920]: Remove up-to-date check from all tasks that extend
 `NpmInstallTask`.
 
+## 2.0.0 - 2017-02-23
+
+### Fixed
+- [LPS-69920]: Fix duplicated NPM arguments while retrying `npm install` in case
+of failure.
+- [LPS-70870]: Fix Node.js download with authenticated proxies.
+
 [LPS-66906]: https://issues.liferay.com/browse/LPS-66906
 [LPS-67023]: https://issues.liferay.com/browse/LPS-67023
 [LPS-67573]: https://issues.liferay.com/browse/LPS-67573
@@ -86,3 +93,4 @@ and the `npm-shrinkwrap.json` file from the project, if present.
 [LPS-69677]: https://issues.liferay.com/browse/LPS-69677
 [LPS-69802]: https://issues.liferay.com/browse/LPS-69802
 [LPS-69920]: https://issues.liferay.com/browse/LPS-69920
+[LPS-70870]: https://issues.liferay.com/browse/LPS-70870