+++ modules/sdk/gradle-plugins/CHANGELOG.markdown
@@ -1111,13 +1111,31 @@ to version 1.0.87.
 - [LPS-70677]: Update the [Liferay Gradle Plugins Jasper JSPC] dependency to
 version 2.0.0.
 
-## 3.2.5 - 2017-02-16
+## 3.2.6 - 2017-02-17
 
 ### Changed
 - [LPS-70707]: Update the [Liferay Gradle Plugins Source Formatter] dependency
 to version 1.0.88.
 - [LPS-70707]: Update the [Liferay Source Formatter] dependency to version
 1.0.355.
+
+## 3.2.7 - 2017-02-22
+
+### Changed
+- [LPS-66853]: Update the [Liferay Gradle Plugins Source Formatter] dependency
+to version 1.0.89.
+- [LPS-66853]: Update the [Liferay Source Formatter] dependency to version
+1.0.356.
+
+## 3.2.8 - 2017-02-24
+
+### Changed
+- [LPS-70870]: Update the [Liferay Gradle Plugins Gulp] dependency to version
+2.0.10.
+- [LPS-70870]: Update the [Liferay Gradle Plugins JS Module Config Generator]
+dependency to version 2.1.8.
+- [LPS-70870]: Update the [Liferay Gradle Plugins JS Transpiler] dependency to
+version 2.3.7.
 
 [Gradle Bundle Plugin]: https://github.com/TomDmitriev/gradle-bundle-plugin
 [Liferay Ant BND]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/ant-bnd
@@ -1225,4 +1243,5 @@ to version 1.0.88.
 [LPS-70618]: https://issues.liferay.com/browse/LPS-70618
 [LPS-70677]: https://issues.liferay.com/browse/LPS-70677
 [LPS-70707]: https://issues.liferay.com/browse/LPS-70707
+[LPS-70870]: https://issues.liferay.com/browse/LPS-70870
 [LRDOCS-3023]: https://issues.liferay.com/browse/LRDOCS-3023